added unified error handling (#205)

+++ backend/src/app/auctions/auctions.controller.ts
@@ -7,12 +7,15 @@ import {
   Param,
   Delete,
   Query,
+  HttpException,
+  HttpStatus,
 } from '@nestjs/common';
 import { AuctionsService } from './auctions.service';
 import { CreateAuctionDto } from './dto/create-auction.dto';
 import { UpdateAuctionDto } from './dto/update-auction.dto';
 import { AuctionDto } from './dto/auction.dto';
 import { AuctionsListParamsDto } from './dto/auctions-list-params.dto';
+import { parseStringError } from '../../common/utils/errors';
 
 @Controller('auctions')
 export class AuctionsController {
@@ -34,19 +37,37 @@ export class AuctionsController {
 
   @Get(':id')
   async findOne(@Param('id') id: string): Promise<AuctionDto> {
-    const auction = await this.auctionsService.findOne(id);
+    const auction = await this.auctionsService.findOne(id).catch(() => null);
+    if (!auction) {
+      throw new HttpException(
+        parseStringError('Auction not found'),
+        HttpStatus.NOT_FOUND,
+      );
+    }
     return AuctionsService.parse(auction);
   }
 
   @Patch(':id')
-  update(@Param('id') id: string, @Body() updateAuctionDto: UpdateAuctionDto) {
+  async update(@Param('id') id: string, @Body() updateAuctionDto: UpdateAuctionDto) {
+    const auction = await this.auctionsService.findOne(id).catch(() => null);
+    if (!auction) {
+      throw new HttpException(
+        parseStringError('Auction not found'),
+        HttpStatus.NOT_FOUND,
+      );
+    }
     return this.auctionsService.update(id, updateAuctionDto);
   }
 
   @Delete(':id')
   async remove(@Param('id') id: string) {
-    const result = await this.auctionsService.remove(id);
-
-    return result;
+    const auction = await this.auctionsService.findOne(id).catch(() => null);
+    if (!auction) {
+      throw new HttpException(
+        parseStringError('Auction not found'),
+        HttpStatus.NOT_FOUND,
+      );
+    }
+    return await this.auctionsService.remove(id);
   }
 }
